+++ typescript/packages/mechanisms/evm/src/upto/client/scheme.ts
@@ -9,7 +9,7 @@ import { createUptoPermit2Payload } from "./permit2";
 import {
   trySignEip2612PermitExtension,
   trySignErc20ApprovalExtension,
-} from "../../shared/permit2-helpers";
+} from "../../shared/extensions";
 import { UptoEvmSchemeOptions } from "./rpc";
 
 /**
